[gojgho-ecommerce-service] user 조회 querydsl 적용 및 불필요한 모듈구조 삭제

+++ service-domain/src/main/resources/db/migration/V1__create_user_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE user (
-                      user_seq INT AUTO_INCREMENT PRIMARY KEY,
+                      user_seq BIGINT AUTO_INCREMENT PRIMARY KEY,
                       user_name VARCHAR(50) NOT NULL,
                       user_email VARCHAR(50) NOT NULL
 );
